requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.4
distlib==0.4.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.19.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.13
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
kombu==5.5.4
nodeenv==1.9.1
openai==1.107.2
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
prompt_toolkit==3.0.52
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.4
pypdfium2==4.30.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tiktoken==0.11.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
virtualenv==20.34.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
